+++ lib/fun_with_flags/store/cache.ex
@@ -1,5 +1,14 @@
 defmodule FunWithFlags.Store.Cache do
-  @moduledoc false
+  @moduledoc """
+  The in-memory cache for the feature flag, backed by an ETS table.
+
+  This module is not meant to be used directly, but some of its functions can be
+  useful to debug flag state.
+  """
+
+  @type ttl :: integer
+  @type cached_at :: integer
+
   use GenServer
   alias FunWithFlags.Timestamps
   alias FunWithFlags.Flag
@@ -11,6 +20,7 @@ defmodule FunWithFlags.Store.Cache do
   ]
 
 
+  @doc false
   def worker_spec do
     if FunWithFlags.Config.cache? do
       %{
@@ -23,6 +33,7 @@ defmodule FunWithFlags.Store.Cache do
   end
 
 
+  @doc false
   def start_link do
     GenServer.start_link(__MODULE__, :ok, [name: __MODULE__])
   end
@@ -31,6 +42,7 @@ defmodule FunWithFlags.Store.Cache do
   # We lookup without going through the GenServer
   # for concurrency and perfomance.
   #
+  @doc false
   def get(flag_name) do
     case :ets.lookup(@table_name, flag_name) do
       [{^flag_name, {flag, timestamp, ttl}}] ->
@@ -39,6 +51,7 @@ defmodule FunWithFlags.Store.Cache do
         {:miss, :not_found, nil}
     end
   end
+
 
   defp validate(name, flag = %Flag{name: name}, timestamp, ttl) do
     if Timestamps.expired?(timestamp, ttl) do
@@ -55,23 +68,36 @@ defmodule FunWithFlags.Store.Cache do
   # We want to always write serially through the
   # GenServer to avoid race conditions.
   #
+  @doc false
   def put(flag = %Flag{}) do
     GenServer.call(__MODULE__, {:put, flag})
   end
 
 
+  @doc """
+  Clears the cache. It will be rebuilt gradually as the public interface of the
+  package is queried.
+  """
+  @spec flush() :: true
   def flush do
     GenServer.call(__MODULE__, :flush)
   end
 
+
+  @doc """
+  Returns the contents of the cache ETS table, for inspection.
+  """
+  @spec dump() :: [{atom, {FunWithFlags.Flag.t, cached_at, ttl}}]
   def dump do
     :ets.tab2list(@table_name)
   end
+
 
   # ------------------------------------------------------------
   # GenServer callbacks
 
 
+  @doc false
   def init(:ok) do
     tab_name = @table_name
     ^tab_name = :ets.new(@table_name, @table_options)
@@ -79,6 +105,7 @@ defmodule FunWithFlags.Store.Cache do
   end
 
 
+  @doc false
   def handle_call({:put, flag = %Flag{name: name}}, _from, state = %{ttl: ttl}) do
     # writing to an ETS table will either return true or raise
     :ets.insert(@table_name, {name, {flag, Timestamps.now, ttl}})
@@ -86,6 +113,7 @@ defmodule FunWithFlags.Store.Cache do
   end
 
 
+  @doc false
   def handle_call(:flush, _from, state) do
     {:reply, :ets.delete_all_objects(@table_name), state}
   end